Move header files under ocispec/
