+++ src/ocispec/read-file.c
@@ -16,7 +16,7 @@
 
 #include <config.h>
 
-#include "read-file.h"
+#include "ocispec/read-file.h"
 
 #include <sys/stat.h>
 #include <stdio.h>
