+++ tests/test-8.c
@@ -20,7 +20,7 @@ along with libocispec.  If not, see <http://www.gnu.org/licenses/>.
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "image_manifest_items_image_manifest_items_schema.h"
+#include "ocispec/image_manifest_items_image_manifest_items_schema.h"
 
 
 int
